modify folder

+++ include/caffe/layers/proposal_target_layer.hpp
@@ -6,7 +6,7 @@
 #include "caffe/blob.hpp"
 #include "caffe/layer.hpp"
 #include "caffe/proto/caffe.pb.h"
-#include "caffe/util/frcnn_util.hpp"
+#include "caffe/util/frcnn_utils.hpp"
 
 namespace caffe {
 
@@ -52,13 +52,13 @@ class ProposalTargetLayer : public Layer<Dtype> {
   virtual void Backward_gpu(const vector<Blob<Dtype>*>& top,
                             const vector<bool>& propagate_down,
                             const vector<Blob<Dtype>*>& bottom) {}
-  void _sample_rois(const vector<Point4f<Dtype> >& all_rois,
-                    const vector<Point4f<Dtype> >& gt_boxes,
+  void _sample_rois(const vector<Point4d<Dtype> >& all_rois,
+                    const vector<Point4d<Dtype> >& gt_boxes,
                     const vector<int>& gt_label, const int fg_rois_per_image,
                     const int rois_per_image, vector<int>& labels,
-                    vector<Point4f<Dtype> >& rois,
-                    vector<vector<Point4f<Dtype> > >& bbox_targets,
-                    vector<vector<Point4f<Dtype> > >& bbox_inside_weights);
+                    vector<Point4d<Dtype> >& rois,
+                    vector<vector<Point4d<Dtype> > >& bbox_targets,
+                    vector<vector<Point4d<Dtype> > >& bbox_inside_weights);
   int n_classes_;
   shared_ptr<Caffe::RNG> rng_;
   int count_;
